TOT/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
sympy>=1.12
openai>=0.27.0
backoff>=2.2.0
python-dotenv==0.9.9

